svc-empleados/requirements.txt
alembic==1.16.4
blinker==1.9.0
click==8.2.1
colorama==0.4.6
et_xmlfile==2.0.0
fire==0.7.0
Flask==3.1.1
Flask-JWT-Extended==4.7.1
flask-marshmallow==1.3.0
Flask-SQLAlchemy==3.1.1
fonttools==4.58.5
gunicorn==23.0.0
greenlet==3.2.4
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
ldap3==2.9.1
lxml==6.0.0
Mako==1.3.10
MarkupSafe==3.0.2
marshmallow==4.0.0
marshmallow-sqlalchemy==1.4.2
numpy==2.2.6
opencv-python-headless==4.12.0.88
openpyxl==3.1.5
pandas==2.3.2
pdf2docx==0.5.8
pyasn1==0.6.1
PyJWT==2.10.1
PyMuPDF==1.26.3
pymysql==1.1.2
python-dateutil==2.9.0.post0
python-docx==1.2.0
python-dotenv==1.1.1
pytest==8.4.2
pytest-flask==1.3.0
pytz==2025.2
six==1.17.0
SQLAlchemy==2.0.43
termcolor==3.1.0
typing_extensions==4.14.1
tzdata==2025.2
Werkzeug==3.1.3